x02-BackEnd/x0201-fastAPI/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
passlib[bcrypt]
python-jose[cryptography]
python-multipart
email-validator
python-dotenv
